+++ framework/Source/CPTAnimationPeriod.m
@@ -440,7 +440,7 @@
 
 -(NSString *)description
 {
-    return [NSString stringWithFormat:@"<%@ from: %@; to: %@; duration: %g, delay: %g>", [super description], self.startValue, self.endValue, self.duration, self.delay];
+    return [NSString stringWithFormat:@"<%@ from: %@; to: %@; duration: %g, delay: %g>", [super description], self.startValue, self.endValue, (double)self.duration, (double)self.delay];
 }
 
 /// @endcond
